Initialize markdown configurations and migrate JS and CSS files
The commit introduces 'kramdown' for markdown rendering in the `_config.yml`, and adds new layout and 'About' pages. It makes changes to the `index.md` by removing the test message and adding a welcome greeting. The JavaScript and CSS files have been relocated to the 'assets' directory for better organization.