Add navigation menu and update project settings
Added a new navigation menu to the website layout and updated project settings. The navigation menu includes a dropdown and it was created in a new HTML file, included in the default layout. Set the spell check inspection tool not to process code in the project's default inspection profile, and added 'US/Central' timezone in the website configuration.