Add tag index, RSS feed, and update _config and head files
Several updates are made to improve site structure and functionalities. _config file is updated to exclude certain files from the production site. A new 'Tag Index' page is created for easier navigation. A new RSS feed is also added for subscribers. In addition, the links in the head file have been updated to be more adaptable.