Refactor post display with include statement
The post display has been refactored to use an include statement instead of directly writing HTML in the "for" loop. A new file "post.html" has been created for this purpose, containing the post structure and elements. The "post.html" script has been included in the "posts.html" and "_layouts/post.html" files for consistent post display across the site.