Update site structure and add new blog post
This commit involves a major restructuring of the website involving updating of CSS files for better post readability, modifications to configuration for improved site navigation, and applying changes to ensure proper date rendering in posts. It also includes the addition of a new blog post announcing the Autonomais open-source library. Various files have been moved to '_