Add estimated read time and deep link for posts
The post inclusion has been modified to include the estimated reading time and a "Go deeper" link for posts that are not marked as 'full'. A new 'read_time.html' component was created to calculate the reading time based on word count. In addition, a 'tags' property has been added to posts to categorize them.