Update layout, rearrange content and add projects section
This commit updates the home layout to modify the container size. The post title structure in the post.html file is rearranged for better readability. The unneeded 'docs_list_title' from menu.yaml is removed. A new projects.yaml file is added to include projects info. The index.markdown file is also greatly updated to include a new projects section presenting contents from the added projects.yaml, and the arrangement of the recent writing section is improved.