Update condition for "Keep reading" button
The condition for displaying the "Keep reading" button has been updated. Specifically, the logic has been adjusted to be less restrictive. Instead of requiring both `include.full` to be false and `post.content` to contain `site.excerpt_separator`, the new logic displays the button when either of these conditions is met.